+++ src/store/reducers/group.mjs
@@ -160,6 +160,9 @@ export function moveGroup( state, { source_data, target_data } ) {
       source_window1.active_tab_id = source_window_active_group1.active_tab_id
       source_window1.highlighted_tab_ids = [ source_window_active_group1.active_tab_id ]
     }
+    if( ! source_window1.active_tab_id ) {
+      source_window1.remove = true
+    }
   }
 
   const target_tab_groups = [ ...windows[ target_window_index ].tab_groups ]