+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Fixed
 - validation error on close last tab in window
-- validation error on moving the last group from a window
+- validation error on moving the last tabs from a window
 - handling when active group is closed
 
 ## [0.0.16] - 2020-02-17